Rebase and update to work with current chain

+++ pkg/gridtypes/deployment.go
@@ -39,7 +39,7 @@ type KeyGetter interface {
 type Deployment struct {
 	Version              int                  `json:"version"`
 	TwinID               uint32               `json:"twin_id"`
-	ContractID           uint32               `json:"contract_id"`
+	ContractID           uint64               `json:"contract_id"`
 	Metadata             string               `json:"metadata"`
 	Description          string               `json:"description"`
 	Expiration           Timestamp            `json:"expiration"`
